refactor: move implementation to design system
