+++ ui/build.gradle.kts
@@ -56,13 +56,6 @@ dependencies {
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-    implementation(libs.androidx.material3)
-    implementation(libs.androidx.ui)
-    implementation(libs.androidx.ui.tooling.preview)
-    implementation(libs.androidx.activity.compose)
-    implementation(libs.androidx.foundation.layout.android)
-    implementation(platform(libs.androidx.compose.bom))
-    implementation(libs.androidx.material3.android)
     implementation(project(":design_system"))
     implementation(project(":viewmodel"))
 }
